requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
feedparser==6.0.11
groq==0.11.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.8.2
requests==2.32.3
